gameplay_service/gameplay/requirements.txt
django
channels
channels-redis
celery
daphne
